requirements.txt
# Ingestion
finnhub-python==2.4.20
kafka-python==2.0.2
requests==2.31.0
python-dotenv==1.0.1

# Model / Inference
torch==2.2.2
transformers==4.40.0
datasets==2.19.0
scikit-learn==1.4.2
mlflow==2.12.2
psycopg2-binary==2.9.9

# Signal / Backtest
yfinance==0.2.38
pandas==2.2.2
numpy==1.26.4

# API
fastapi==0.111.0
uvicorn[standard]==0.29.0
